PostItemとBorderLineをファイル分割した
