+++ frontend/src/shared/components/BorderLine.tsx
@@ -1,0 +1,7 @@
+export function BorderLine() {
+    return(
+    <div className='flex items-center justify-center py-3'>
+      <hr className="w-full border-border" />
+    </div>
+    );
+}
